third_party/WebKit/LayoutTests/fast/dom/shadow/get-destination-insertion-points-skips-user-agent-shadow-expected.txt

Tests that getDestinationInsertionPoints() does not leak insertion points in user agent shadow tree

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.getDestinationInsertionPoints().length is 0
PASS successfullyParsed is true

TEST COMPLETE
